feat: Menu ordering function in shopping cart

+++ src/main/java/com/dineq/dineqbe/repository/TableOrderRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TableOrderRepository extends JpaRepository<TableOrderEntity, Integer> {
+public interface TableOrderRepository extends JpaRepository<TableOrderEntity, Long> {
 }
